requirements.txt
django=1.10.2
djangorestframework=3.5.1
pygments=2.1.3
npm=0.1.1
django-webpack-loader=0.3.3